requirements.txt
streamlit==1.44.1
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
plotly==6.0.1
python-docx==1.1.2
openai==1.75.0
pypandoc==1.15
python-dotenv==1.1.0
